# improve startup speed for expanded probes w/o *