Improved speed by not performing wildcard match on probes without a
wildcard. Speedup of complex scripts (especially those using tracepoints
with arguments) will be much faster now. As an example from our toos:

Before:

```
# time bpftrace -d statsnoop.bt
...
bpftrace -d statsnoop.bt  1.67s user 0.06s system 99% cpu 1.748 total
```

After:

```
# time bpftrace -d statsnoop.bt
...
bpftrace -d statsnoop.bt  0.06s user 0.02s system 80% cpu 0.098 total
```

Ref: https://github.com/iovisor/bpftrace/issues/526